App deployment

+++ src/main/java/com/test/restapi/database/DatabaseClass.java
@@ -7,10 +7,10 @@ import com.test.restapi.model.Message;
 
 public class DatabaseClass {
 
-	private static Map<Long, Message> messages = new HashMap<>();
+	private static Map<Integer, Message> messages = new HashMap<>();
 //	private static Map<Long, Message> profile = new HashMap<>();
 	
-	public static Map<Long, Message> getMessages() {
+	public static Map<Integer, Message> getMessages() {
 		return messages;
 	}
